feat: add basic logging system
Add a logging system with macros that expand to nothing on release
builds and to proper log function calls on debug (development) builds.

Add basic log file handling: the system creates a log file and prints
log messages to it. Might not work properly with the shared modules at
this moment.